tests/submission/W3C/domhtml/HTMLObjectElement06.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-object-element.html#dom-object-data" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<object data="/support/html/sample.html" height="80" type="text/html" width="550"></object>
<script>
var domtest = async_test("The IDL attribute data must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vdata;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
      testNode = nodeList.item(0);
      vdata = testNode.data;

      assert_uri_equals(document.location.protocol,null,document.location.host,"sample.html",null,null,null,true,vdata);
       
    domtest.done();
}));

</script>
<p>The IDL attribute data must reflect the respective content attribute of the same name.</p>
<div id="log"></div>
</body></html>
